Investigated the regular 5 seconds hangups.
There is a problem with NSSwitch, where a DNS resolution is made to
the DNS servers before checking the prax resolver. It creates an
overhead between half a second and 5 seconds on my setup.

Moving the prax resolver before the dns one resolves this problem.
Yet it may cause DNS resolutions bugs.