Fix: refill the pool of threads.
The pool of thread was static, and wasn't refilled automatically
when a thread exited, thus potentially leading to no thread to
execute the requests (both in Prax::Server and Racker::Server).

Prax::Workers aims to fix the problem by running the thread block
inside a begin/rescue statement which will spawn a fresh thread.