Removed support for specific version managers
This is an attempt at skipping support for specific version
managers as per #36, while still permitting the user to choose
the ruby version under which racker must run.

It works by changing the call to Process.spawn to a string instead
of a list of argument, which forces the command to be run through a
shell and not directly from the ruby process. We also clean the
environment from everything except for PATH which is set to
ORIG_PATH, which is the PATH before Prax is started, thus skipping
any ruby version changes.

It works perfectly with rbenv.